requirement.txt
click
cryptography<39,>=38.0.0
snowflake-connector-python[pandas]